Added integration test for TasksController and fixed memord db injection problem.

+++ SimpleTaskSystem-Core/src/Acme.SimpleTaskApp.Web/Models/Tasks/IndexViewModel.cs
@@ -49,7 +49,7 @@ namespace Acme.SimpleTaskApp.Web.Models.Tasks
                         new SelectListItem
                         {
                             Text = localizationManager.GetString(SimpleTaskAppConsts.LocalizationSourceName, $"TaskState_{state}"),
-                            Value = ((byte)state).ToString(),
+                            Value = state.ToString(),
                             Selected = state == SelectedTaskState
                         })
             );
